[refactor] Moved 'getRandomInt' to its own TS file
