+++ src/CompanyManagement.js
@@ -7,7 +7,7 @@ import {Player}                                         from "./Player";
 
 import Decimal                                          from "decimal.js";
 import {dialogBoxCreate}                                from "../utils/DialogBox";
-import {getRandomInt, removeElementById,
+import {removeElementById,
         createElement, createAccordionElement,
         removeChildrenFromElement, createPopup,
         clearSelector}                                  from "../utils/HelperFunctions";
@@ -15,6 +15,7 @@ import {Reviver, Generic_toJSON,
         Generic_fromJSON}                               from "../utils/JSONReviver";
 import numeral                                          from "numeral/min/numeral.min";
 import {formatNumber, generateRandomString}             from "../utils/StringHelperFunctions";
+import {getRandomInt}                                   from "../utils/helpers/getRandomInt";
 import {isString}                                       from "../utils/helpers/isString";
 import {yesNoBoxCreate, yesNoTxtInpBoxCreate,
         yesNoBoxGetYesButton, yesNoBoxGetNoButton,
