+++ src/Infiltration.js
@@ -3,7 +3,8 @@ import {CONSTANTS}                              from "./Constants";
 import {Engine}                                 from "./engine";
 import {Player}                                 from "./Player";
 import {dialogBoxCreate}                        from "../utils/DialogBox";
-import {clearEventListeners, getRandomInt}      from "../utils/HelperFunctions";
+import {clearEventListeners}                    from "../utils/HelperFunctions";
+import {getRandomInt}                           from "../utils/helpers/getRandomInt";
 import {infiltrationBoxCreate}                  from "../utils/InfiltrationBox";
 import {formatNumber}                           from "../utils/StringHelperFunctions";
