+++ src/Missions.js
@@ -3,10 +3,11 @@ import {Engine}                                     from "./engine";
 import {displayFactionContent}                      from "./Faction";
 import {Player}                                     from "./Player";
 import {dialogBoxCreate}                            from "../utils/DialogBox";
-import {addOffset, getRandomInt,
+import {addOffset,
         clearEventListenersEl,
         clearEventListeners}                        from "../utils/HelperFunctions";
 import {formatNumber}                               from "../utils/StringHelperFunctions";
+import {getRandomInt}                               from "../utils/helpers/getRandomInt";
 import {isString}                                   from "../utils/helpers/isString";
 import jsplumb                                      from 'jsplumb'
